Add support for pages

+++ MyApp/Configure.Ssg.cs
@@ -10,20 +10,23 @@ public class ConfigureSsg : IHostingStartup
         .ConfigureServices(services =>
         {
             services.AddSingleton<RazorPagesEngine>();
-            services.AddSingleton<Blog>();
+            services.AddSingleton<MarkdownPages>();
             services.AddSingleton<WhatsNew>();
+            services.AddSingleton<Blog>();
         })
         .ConfigureAppHost(
             appHost => appHost.Plugins.Add(new CleanUrlsFeature()),
             afterPluginsLoaded: appHost =>
             {
+                var markdownPages = appHost.Resolve<MarkdownPages>();
                 var whatsNew = appHost.Resolve<WhatsNew>();
-                whatsNew.VirtualFiles = appHost.VirtualFiles;
-                whatsNew.LoadFeatures("_whatsnew");
-                
                 var blogPosts = appHost.Resolve<Blog>();
-                blogPosts.VirtualFiles = appHost.VirtualFiles;
-                blogPosts.LoadPosts("_posts");
+                new MarkdownPagesBase[] { markdownPages, whatsNew, blogPosts }
+                    .Each(x => x.VirtualFiles = appHost.VirtualFiles);
+
+                markdownPages.LoadFrom("_pages");
+                whatsNew.LoadFrom("_whatsnew");
+                blogPosts.LoadFrom("_posts");
             },
             afterAppHostInit: appHost =>
             {
@@ -31,7 +34,8 @@ public class ConfigureSsg : IHostingStartup
                 AppTasks.Register("prerender", args =>
                 {
                     var distDir = appHost.ContentRootDirectory.RealPath.CombineWith("dist");
-                    FileSystemVirtualFiles.DeleteDirectory(distDir);
+                    if (Directory.Exists(distDir))
+                        FileSystemVirtualFiles.DeleteDirectory(distDir);
                     FileSystemVirtualFiles.CopyAll(
                         new DirectoryInfo(appHost.ContentRootDirectory.RealPath.CombineWith("wwwroot")),
                         new DirectoryInfo(distDir));
@@ -39,25 +43,4 @@ public class ConfigureSsg : IHostingStartup
                     RazorSsg.PrerenderAsync(appHost, razorFiles, distDir).GetAwaiter().GetResult();
                 });
             });
-
-}
-
-public class MarkdownFileInfo
-{
-    public string Path { get; set; } = default!;
-    public string? Slug { get; set; }
-    public string? FileName { get; set; }
-    public string? HtmlFileName { get; set; }
-    public string? Title { get; set; }
-    public string? Summary { get; set; }
-    public string? Image { get; set; }
-    public string? Author { get; set; }
-    public List<string> Tags { get; set; } = new();
-    public DateTime? Date { get; set; }
-    public string? Content { get; set; }
-    public string? Url { get; set; }
-    public string? Preview { get; set; }
-    public string? HtmlPage { get; set; }
-    public int? WordCount { get; set; }
-    public int? LineCount { get; set; }
 }
